Test3/requirements.txt
gradio
transformers
sentence-transformers
faiss-cpu
pymupdf
tabula-py
pillow
torch